+++ lib/src/test/java/nl/knaw/dans/schema/ValidXmlsTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.dansschema;
+package nl.knaw.dans.schema;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
