+++ lib/src/test/java/nl/knaw/dans/schema/XmlReader.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.dansschema;
+package nl.knaw.dans.schema;
 
 import org.w3c.dom.Document;
 import org.xml.sax.InputSource;
